feat(sheets-advanced): support chart (#6)

+++ common/shared/vite/prepend-umd-raw-plugin.ts
@@ -64,8 +64,12 @@ export function prependUMDRawPlugin(options: IOptions): Plugin {
                 });
 
                 const __locale = path.resolve(__localeDir, `${localeKey}.js`);
-                fs.ensureFileSync(__locale);
-                fs.writeFileSync(__locale, localeContents.join('\n\n'));
+
+                if (fs.existsSync(__locale)) {
+                    localeContents.push(fs.readFileSync(__locale, 'utf8'));
+
+                    fs.writeFileSync(__locale, localeContents.join('\n\n'));
+                }
             });
         },
     };
